requirements.txt
appnope==0.1.0
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.17.0
backports.shutil-get-terminal-size==1.0.0
channels==0.17.3
constantly==15.1.0
daphne==0.15.0
decorator==4.0.10
Django==1.10.4
django-extensions==1.7.5
enum34==1.1.6
haikunator==2.1.0
incremental==16.10.1
ipython==5.1.0
ipython-genutils==0.1.0
msgpack-python==0.4.8
mysqlclient==1.3.9
olefile==0.44
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
python-decouple==3.0
redis==2.10.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
Twisted==16.6.0
txaio==2.5.2
Unipath==1.1
wcwidth==0.1.7
zope.interface==4.3.3
